added auth_config and manager files
