+++ backend/app/api/v1/routers/step_router.py
@@ -1,5 +1,4 @@
 from app.courses.models import Step
-from app.users.dependencies import current_user
 from app.courses.dependencies import StepsServiceDep
 from app.courses.schemas import StepCreate, StepUpdate
 from app.core.router import ParentItemRouterWithUser
