+++ backend/app/api/v1/routers/submission_router.py
@@ -3,7 +3,7 @@ from app.submissions.dependencies import SubmissionsServiceDep
 from app.submissions.models import Submission
 from app.submissions.schemas import SubmissionCreate
 
-from app.users.dependencies import current_user
+from app.users.auth_config import current_user
 from app.users.models import User
 
 from app.utils.exceptions import OpenAPIDocExtraResponse
