+++ backend/app/users/auth_config.py
@@ -1,0 +1,34 @@
+from pydantic import UUID4
+
+from fastapi_users import FastAPIUsers
+from fastapi_users.authentication import (
+    AuthenticationBackend,
+    BearerTransport,
+    JWTStrategy,
+)
+
+from app.core.config import settings
+from app.users.models import User
+from app.users.manager import get_user_manager
+
+
+bearer_transport = BearerTransport(tokenUrl="v1/auth/jwt/login")
+
+def get_jwt_strategy() -> JWTStrategy:
+    """Defining the strategy: storing the token in the form of a JWT."""
+    return JWTStrategy(secret=settings.secret, lifetime_seconds=3600)
+
+
+auth_backend = AuthenticationBackend(
+    name="jwt",
+    transport=bearer_transport,
+    get_strategy=get_jwt_strategy,
+)
+
+fastapi_users = FastAPIUsers[User, UUID4](
+    get_user_manager,
+    [auth_backend],
+)
+
+current_user = fastapi_users.current_user(active=True)
+current_superuser = fastapi_users.current_user(active=True, superuser=True)
