+++ backend/app/users/dependencies.py
@@ -1,78 +1,11 @@
-from pydantic import UUID4
-from typing import Annotated, Optional, Union
+from typing import Annotated
 
-from fastapi import Depends, Request
-from fastapi_users import (
-    BaseUserManager,
-    FastAPIUsers,
-    UUIDIDMixin,
-    InvalidPasswordException,
-)
-from fastapi_users_db_sqlmodel import SQLModelUserDatabaseAsync
-from fastapi_users.authentication import (
-    AuthenticationBackend,
-    BearerTransport,
-    JWTStrategy,
-)
-from sqlmodel.ext.asyncio.session import AsyncSession
+from fastapi import Depends
 
-from app.core.config import settings
-from app.core.db import get_async_session
 from app.core.transaction_manager import TManagerDep
-from app.users.models import User, UserProgress
-from app.users.schemas import UserCreate
+from app.users.models import  UserProgress
 from app.users.service import UserProgressService
-
-
-async def get_user_db(session: AsyncSession = Depends(get_async_session)):
-    yield SQLModelUserDatabaseAsync(session=session, user_model=User)
-
-
-bearer_transport = BearerTransport(tokenUrl="auth/jwt/login")
-
-
-def get_jwt_strategy() -> JWTStrategy:
-    """Defining the strategy: storing the token in the form of a JWT."""
-    return JWTStrategy(secret=settings.secret, lifetime_seconds=3600)
-
-
-# Creating an authentication backend object with the selected parameters.
-auth_backend = AuthenticationBackend(
-    name="jwt",
-    transport=bearer_transport,
-    get_strategy=get_jwt_strategy,
-)
-
-
-class UserManager(UUIDIDMixin, BaseUserManager[User, UUID4]):
-    async def validate_password(
-        self,
-        password: str,
-        user: Union[UserCreate, User],
-    ) -> None:
-        if len(password) < 3:
-            raise InvalidPasswordException(
-                reason="Password should be at least 3 characters"
-            )
-        if user.email in password:
-            raise InvalidPasswordException(reason="Password should not contain e-mail")
-
-    async def on_after_register(self, user: User, request: Optional[Request] = None):
-        """Method for actions after successful user registration."""
-        print(f"The user {user.email} is registered.")
-
-
-async def get_user_manager(user_db=Depends(get_user_db)):
-    yield UserManager(user_db)
-
-
-fastapi_users = FastAPIUsers[User, UUID4](
-    get_user_manager,
-    [auth_backend],
-)
-
-current_user = fastapi_users.current_user(active=True)
-current_superuser = fastapi_users.current_user(active=True, superuser=True)
+from app.users.manager import UserManager, get_user_manager
 
 
 def get_user_progress_service(
